+++ app/src/main/java/AcquireProject/Merger.java
@@ -21,14 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  *
- * Created by Emily Elzinga
  *
  * 11/5/2021
  *
  * performs all the functions and steps when merging two hotel chains.
  *
  * @since 1.0.0
- * @author Michael Collier
+ * @author Michael Collier, Emily Elzinga
  */
